requirements.txt
Pillow==6.1.0
pytest==6.0.1
pytest-html==2.1.1
pytest-xdist==2.1.0